third_party/WebKit/LayoutTests/fast/dom/time-ranges-exceptions.html

<!DOCTYPE html>
<html>
<head>
    <script src="../../resources/js-test.js"></script>
</head>
<body>
    <script>
        description("Verify that TextRanges exceptions are properly messaged to developers.");
        
        var el = document.createElement('video');
        var timeRanges = el.buffered;

        shouldThrow('el.buffered.start(1)');
        shouldThrow('el.buffered.end(1)');
    </script>
</body>
</html>
