third_party/WebKit/LayoutTests/fast/dom/time-ranges-exceptions-expected.txt

Verify that TextRanges exceptions are properly messaged to developers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS el.buffered.start(1) threw exception IndexSizeError: Failed to execute 'start' on 'TimeRanges': The index provided (1) is not less than the object's length (0)..
PASS el.buffered.end(1) threw exception IndexSizeError: Failed to execute 'end' on 'TimeRanges': The index provided (1) is not less than the object's length (0)..
PASS successfullyParsed is true

TEST COMPLETE
